+++ bitbybit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## bitbybit 2.0.0
+
+This version expects arbitrary-int 2.x.
+
+### Added
+
+- Support for signed arbitrary-int integers as field types, e.g. `i3`, `i24`, etc.
+
 ## bitbybit 1.4.0
 
 This is the final version to support arbitrary-int 1.x. Future versions will require arbitrary-int 2.x.